Remove hte FixMultipleContext type.
