+++ src/EditorFeatures/Test/Diagnostics/AbstractUserDiagnosticTest.cs
@@ -189,7 +189,7 @@ namespace Microsoft.CodeAnalysis.Editor.UnitTests.Diagnostics
             {
                 // Bulk fixing diagnostics in selected scope.                    
                 var diagnosticsToFix = ImmutableDictionary.CreateRange(SpecializedCollections.SingletonEnumerable(KeyValuePair.Create(document, diagnostics.ToImmutableArray())));
-                return FixMultipleContext.Create(diagnosticsToFix, fixer, fixAllActionId, CancellationToken.None);
+                return FixAllContext.Create(diagnosticsToFix, fixer, fixAllActionId, CancellationToken.None);
             }
 
             var diagnostic = diagnostics.First();
