+++ src/Workspaces/Core/Portable/CodeFixes/FixAllOccurrences/FixAllContext.cs
@@ -314,5 +314,64 @@ namespace Microsoft.CodeAnalysis.CodeFixes
                 .Where(fixAllProviderInfo.CanBeFixed)
                 .Select(d => d.Id);
         }
+
+        #region FixMultiple
+
+        /// <summary>
+        /// Creates a new <see cref="FixAllContext"/>.
+        /// Use this overload when applying fix multiple diagnostics with a source location.
+        /// </summary>
+        /// <param name="diagnosticsToFix">Specific set of diagnostics to fix. Must be a non-empty set.</param>
+        /// <param name="codeFixProvider">Underlying <see cref="CodeFixes.CodeFixProvider"/> which triggered this fix all.</param>
+        /// <param name="codeActionEquivalenceKey">The <see cref="CodeAction.EquivalenceKey"/> value expected of a <see cref="CodeAction"/> participating in this fix all.</param>
+        /// <param name="cancellationToken">Cancellation token for fix all computation.</param>
+        internal static FixAllContext Create(
+            ImmutableDictionary<Document, ImmutableArray<Diagnostic>> diagnosticsToFix,
+            CodeFixProvider codeFixProvider,
+            string codeActionEquivalenceKey,
+            CancellationToken cancellationToken)
+        {
+            var triggerDocument = diagnosticsToFix.First().Key;
+            var diagnosticIds = GetDiagnosticsIds(diagnosticsToFix.Values);
+            var diagnosticProvider = new FixMultipleDiagnosticProvider(diagnosticsToFix);
+            return new FixAllContext(
+                triggerDocument, codeFixProvider, FixAllScope.Custom, codeActionEquivalenceKey, diagnosticIds, diagnosticProvider, cancellationToken);
+        }
+
+        /// <summary>
+        /// Creates a new <see cref="FixAllContext"/>.
+        /// Use this overload when applying fix multiple diagnostics with no source location.
+        /// </summary>
+        /// <param name="diagnosticsToFix">Specific set of diagnostics to fix. Must be a non-empty set.</param>
+        /// <param name="codeFixProvider">Underlying <see cref="CodeFixes.CodeFixProvider"/> which triggered this fix all.</param>
+        /// <param name="codeActionEquivalenceKey">The <see cref="CodeAction.EquivalenceKey"/> value expected of a <see cref="CodeAction"/> participating in this fix all.</param>
+        /// <param name="cancellationToken">Cancellation token for fix all computation.</param>
+        internal static FixAllContext Create(
+            ImmutableDictionary<Project, ImmutableArray<Diagnostic>> diagnosticsToFix,
+            CodeFixProvider codeFixProvider,
+            string codeActionEquivalenceKey,
+            CancellationToken cancellationToken)
+        {
+            var triggerProject = diagnosticsToFix.First().Key;
+            var diagnosticIds = GetDiagnosticsIds(diagnosticsToFix.Values);
+            var diagnosticProvider = new FixMultipleDiagnosticProvider(diagnosticsToFix);
+            return new FixAllContext(triggerProject, codeFixProvider, FixAllScope.Custom, codeActionEquivalenceKey, diagnosticIds, diagnosticProvider, cancellationToken);
+        }
+
+        private static ImmutableHashSet<string> GetDiagnosticsIds(IEnumerable<ImmutableArray<Diagnostic>> diagnosticsCollection)
+        {
+            var uniqueIds = ImmutableHashSet.CreateBuilder<string>();
+            foreach (var diagnostics in diagnosticsCollection)
+            {
+                foreach (var diagnostic in diagnostics)
+                {
+                    uniqueIds.Add(diagnostic.Id);
+                }
+            }
+
+            return uniqueIds.ToImmutable();
+        }
+
+        #endregion
     }
 }
