+++ src/Workspaces/Core/Portable/Workspaces.csproj
@@ -308,8 +308,7 @@
     <Compile Include="CodeFixes\FixAllOccurrences\BatchSimplificationFixAllProvider.cs" />
     <Compile Include="CodeFixes\FixAllOccurrences\FixAllContext.FixAllDiagnosticProvider.cs" />
     <Compile Include="CodeFixes\FixAllOccurrences\FixAllProviderInfo.cs" />
-    <Compile Include="CodeFixes\FixAllOccurrences\FixMultipleContext.DiagnosticProvider.cs" />
-    <Compile Include="CodeFixes\FixAllOccurrences\FixMultipleContext.cs" />
+    <Compile Include="CodeFixes\FixAllOccurrences\FixAllContext.FixMultipleDiagnosticProvider.cs" />
     <Compile Include="CodeFixes\FixAllOccurrences\FixAllLogger.cs" />
     <Compile Include="CodeFixes\FixAllOccurrences\FixAllContext.DiagnosticProvider.cs" />
     <Compile Include="CodeFixes\Supression\ISuppressionFixProvider.cs" />
